requirements.txt
appdirs==1.4.4
coloredlogs==15.0.1
cvxopt==1.3.1
cvxpy==1.3.1
gdown==4.7.1
hydra-core==1.3.2
matplotlib==3.7.1
meshzoo==0.11.1
numpy==1.23.1
numpy==1.24.3
omegaconf==2.3.0
pandas==2.0.1
Pillow==9.5.0
Pillow==10.4.0
pymoo==0.6.0.1
Requests==2.32.3
scikit_learn==1.3.2
scipy==1.14.1
torch==2.0.1
torchmetrics==0.10.3
torchvision==0.15.2
tqdm==4.66.4
tqdm==4.65.0
